requirements.txt
aiogram
asyncio
python-dotenv
segno
pymongo
dateutils
requests